Resolved Build Error

+++ Assets/Scripts/HealthBarScript.cs
@@ -10,15 +10,15 @@ public class HealthBarScript : MonoBehaviour
 	public Gradient gradient;
 	public Image fill;
 
-	public void SetMaxHealth(int health)
+	public void SetMaxHealth(float health)
 	{
 		slider.maxValue = health;
 		slider.value = health;
 
-		fill.color = gradient.Evaluate(1f);
+		fill.color = gradient.Evaluate(1);
 	}
 
-	public void SetHealth(int health)
+	public void SetHealth(float health)
 	{
 		slider.value = health;
